requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch
torchvision
torchaudio
pandas
langchain
langchain-community
langchain-google-genai
sentence-transformers
chromadb
python-dotenv
mcp